+++ api/authorization.pb.go
@@ -6,17 +6,22 @@ package authorization
 import (
 	context "context"
 	fmt "fmt"
+	_ "github.com/gogo/protobuf/gogoproto"
 	proto "github.com/gogo/protobuf/proto"
+	github_com_gogo_protobuf_types "github.com/gogo/protobuf/types"
+	_ "github.com/golang/protobuf/ptypes/timestamp"
 	uuid "github.com/koverto/uuid"
 	grpc "google.golang.org/grpc"
 	io "io"
 	math "math"
+	time "time"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
 var _ = fmt.Errorf
 var _ = math.Inf
+var _ = time.Kitchen
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the proto package it is being compiled against.
@@ -24,22 +29,24 @@ var _ = math.Inf
 // proto package needs to be updated.
 const _ = proto.GoGoProtoPackageIsVersion2 // please upgrade the proto package
 
-type TokenRequest struct {
-	UserID *uuid.UUID `protobuf:"bytes,1,opt,name=userID,proto3" json:"userID,omitempty"`
+type Claims struct {
+	ID        *uuid.UUID `protobuf:"bytes,1,opt,name=ID,proto3" json:"ID,omitempty"`
+	Subject   *uuid.UUID `protobuf:"bytes,2,opt,name=subject,proto3" json:"subject,omitempty"`
+	ExpiresAt *time.Time `protobuf:"bytes,3,opt,name=expiresAt,proto3,stdtime" json:"expiresAt,omitempty"`
 }
 
-func (m *TokenRequest) Reset()         { *m = TokenRequest{} }
-func (m *TokenRequest) String() string { return proto.CompactTextString(m) }
-func (*TokenRequest) ProtoMessage()    {}
-func (*TokenRequest) Descriptor() ([]byte, []int) {
+func (m *Claims) Reset()         { *m = Claims{} }
+func (m *Claims) String() string { return proto.CompactTextString(m) }
+func (*Claims) ProtoMessage()    {}
+func (*Claims) Descriptor() ([]byte, []int) {
 	return fileDescriptor_1dbbe58d1e51a797, []int{0}
 }
-func (m *TokenRequest) XXX_Unmarshal(b []byte) error {
+func (m *Claims) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
 }
-func (m *TokenRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+func (m *Claims) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	if deterministic {
-		return xxx_messageInfo_TokenRequest.Marshal(b, m, deterministic)
+		return xxx_messageInfo_Claims.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
 		n, err := m.MarshalTo(b)
@@ -49,27 +56,41 @@ func (m *TokenRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error)
 		return b[:n], nil
 	}
 }
-func (m *TokenRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TokenRequest.Merge(m, src)
+func (m *Claims) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_Claims.Merge(m, src)
 }
-func (m *TokenRequest) XXX_Size() int {
+func (m *Claims) XXX_Size() int {
 	return m.Size()
 }
-func (m *TokenRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_TokenRequest.DiscardUnknown(m)
+func (m *Claims) XXX_DiscardUnknown() {
+	xxx_messageInfo_Claims.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_TokenRequest proto.InternalMessageInfo
+var xxx_messageInfo_Claims proto.InternalMessageInfo
 
-func (m *TokenRequest) GetUserID() *uuid.UUID {
+func (m *Claims) GetID() *uuid.UUID {
 	if m != nil {
-		return m.UserID
+		return m.ID
+	}
+	return nil
+}
+
+func (m *Claims) GetSubject() *uuid.UUID {
+	if m != nil {
+		return m.Subject
+	}
+	return nil
+}
+
+func (m *Claims) GetExpiresAt() *time.Time {
+	if m != nil {
+		return m.ExpiresAt
 	}
 	return nil
 }
 
 type Token struct {
-	Token string `protobuf:"bytes,1,opt,name=Token,proto3" json:"Token,omitempty"`
+	Token string `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
 }
 
 func (m *Token) Reset()         { *m = Token{} }
@@ -112,85 +133,36 @@ func (m *Token) GetToken() string {
 	return ""
 }
 
-type TokenResponse struct {
-	Success bool       `protobuf:"varint,1,opt,name=success,proto3" json:"success,omitempty"`
-	UserID  *uuid.UUID `protobuf:"bytes,2,opt,name=userID,proto3" json:"userID,omitempty"`
-}
-
-func (m *TokenResponse) Reset()         { *m = TokenResponse{} }
-func (m *TokenResponse) String() string { return proto.CompactTextString(m) }
-func (*TokenResponse) ProtoMessage()    {}
-func (*TokenResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_1dbbe58d1e51a797, []int{2}
-}
-func (m *TokenResponse) XXX_Unmarshal(b []byte) error {
-	return m.Unmarshal(b)
-}
-func (m *TokenResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	if deterministic {
-		return xxx_messageInfo_TokenResponse.Marshal(b, m, deterministic)
-	} else {
-		b = b[:cap(b)]
-		n, err := m.MarshalTo(b)
-		if err != nil {
-			return nil, err
-		}
-		return b[:n], nil
-	}
-}
-func (m *TokenResponse) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TokenResponse.Merge(m, src)
-}
-func (m *TokenResponse) XXX_Size() int {
-	return m.Size()
-}
-func (m *TokenResponse) XXX_DiscardUnknown() {
-	xxx_messageInfo_TokenResponse.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TokenResponse proto.InternalMessageInfo
-
-func (m *TokenResponse) GetSuccess() bool {
-	if m != nil {
-		return m.Success
-	}
-	return false
-}
-
-func (m *TokenResponse) GetUserID() *uuid.UUID {
-	if m != nil {
-		return m.UserID
-	}
-	return nil
-}
-
 func init() {
-	proto.RegisterType((*TokenRequest)(nil), "authorization.TokenRequest")
+	proto.RegisterType((*Claims)(nil), "authorization.Claims")
 	proto.RegisterType((*Token)(nil), "authorization.Token")
-	proto.RegisterType((*TokenResponse)(nil), "authorization.TokenResponse")
 }
 
 func init() { proto.RegisterFile("authorization.proto", fileDescriptor_1dbbe58d1e51a797) }
 
 var fileDescriptor_1dbbe58d1e51a797 = []byte{
-	// 268 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0x12, 0x4e, 0x2c, 0x2d, 0xc9,
-	0xc8, 0x2f, 0xca, 0xac, 0x4a, 0x2c, 0xc9, 0xcc, 0xcf, 0xd3, 0x2b, 0x28, 0xca, 0x2f, 0xc9, 0x17,
-	0xe2, 0x45, 0x11, 0x94, 0x52, 0x4a, 0xcf, 0x2c, 0xc9, 0x28, 0x4d, 0xd2, 0x4b, 0xce, 0xcf, 0xd5,
-	0xcf, 0xce, 0x2f, 0x4b, 0x2d, 0x2a, 0xc9, 0xd7, 0x2f, 0x2d, 0xcd, 0x4c, 0x01, 0x13, 0x10, 0x2d,
-	0x4a, 0x46, 0x5c, 0x3c, 0x21, 0xf9, 0xd9, 0xa9, 0x79, 0x41, 0xa9, 0x85, 0xa5, 0xa9, 0xc5, 0x25,
-	0x42, 0x4a, 0x5c, 0x6c, 0xa5, 0xc5, 0xa9, 0x45, 0x9e, 0x2e, 0x12, 0x8c, 0x0a, 0x8c, 0x1a, 0xdc,
-	0x46, 0x5c, 0x7a, 0x60, 0xc5, 0xa1, 0xa1, 0x9e, 0x2e, 0x41, 0x50, 0x19, 0x25, 0x59, 0x2e, 0x56,
-	0xb0, 0x1e, 0x21, 0x11, 0x28, 0x03, 0xac, 0x96, 0x33, 0x08, 0xc2, 0x51, 0xf2, 0xe5, 0xe2, 0x85,
-	0x1a, 0x59, 0x5c, 0x90, 0x9f, 0x57, 0x9c, 0x2a, 0x24, 0xc1, 0xc5, 0x5e, 0x5c, 0x9a, 0x9c, 0x9c,
-	0x5a, 0x5c, 0x0c, 0x56, 0xc8, 0x11, 0x04, 0xe3, 0x22, 0xd9, 0xc6, 0x84, 0xcb, 0x36, 0xa3, 0x2b,
-	0x8c, 0x5c, 0xbc, 0x8e, 0xc8, 0xfe, 0x12, 0xb2, 0xe5, 0x62, 0x73, 0x2e, 0x4a, 0x4d, 0x2c, 0x49,
-	0x15, 0x92, 0xd6, 0x43, 0x0d, 0x06, 0x64, 0xaf, 0x48, 0x89, 0x60, 0x93, 0x54, 0x62, 0x10, 0x72,
-	0xe0, 0xe2, 0x08, 0x4b, 0xcc, 0xc9, 0x4c, 0x01, 0x19, 0x80, 0x55, 0x8d, 0x94, 0x0c, 0x76, 0x63,
-	0x21, 0xde, 0x51, 0x62, 0x10, 0x72, 0xe2, 0xe2, 0xf2, 0xcc, 0x2b, 0xa3, 0xc8, 0x0c, 0x27, 0x89,
-	0x13, 0x8f, 0xe4, 0x18, 0x2f, 0x3c, 0x92, 0x63, 0x7c, 0xf0, 0x48, 0x8e, 0x71, 0xc2, 0x63, 0x39,
-	0x86, 0x0b, 0x8f, 0xe5, 0x18, 0x6e, 0x3c, 0x96, 0x63, 0x48, 0x62, 0x03, 0xc7, 0x8c, 0x31, 0x20,
-	0x00, 0x00, 0xff, 0xff, 0x9d, 0x7a, 0x52, 0x3f, 0xe3, 0x01, 0x00, 0x00,
+	// 336 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x74, 0x90, 0xcd, 0x4e, 0xc2, 0x40,
+	0x14, 0x85, 0x3b, 0x28, 0x28, 0xd7, 0xb0, 0x19, 0x31, 0x69, 0x9a, 0x58, 0x4c, 0xe3, 0xc2, 0xd5,
+	0x34, 0x80, 0x89, 0x31, 0x31, 0x46, 0x7e, 0x36, 0xdd, 0x12, 0x70, 0xdf, 0xc2, 0x58, 0x46, 0x28,
+	0x97, 0xb4, 0x53, 0x62, 0x7c, 0x04, 0x57, 0xbc, 0x92, 0x3b, 0x97, 0x2c, 0xdd, 0x69, 0xe8, 0x8b,
+	0x18, 0xa6, 0x56, 0x41, 0x61, 0xd3, 0xdc, 0x9e, 0xf9, 0xce, 0x99, 0x33, 0x17, 0x8e, 0xdd, 0x58,
+	0x0e, 0x31, 0x14, 0xcf, 0xae, 0x14, 0x38, 0x61, 0xd3, 0x10, 0x25, 0xd2, 0xd2, 0x86, 0x68, 0x5c,
+	0xfa, 0x42, 0x0e, 0x63, 0x8f, 0xf5, 0x31, 0xb0, 0x7d, 0xf4, 0xd1, 0x56, 0x94, 0x17, 0x3f, 0xdc,
+	0xcd, 0xaa, 0xac, 0xce, 0xaa, 0x4a, 0x54, 0x9a, 0x9a, 0xd2, 0x10, 0xc3, 0x5a, 0x73, 0x8d, 0x70,
+	0xc6, 0x43, 0x89, 0x76, 0x1c, 0x8b, 0x81, 0xfa, 0x7c, 0x33, 0x15, 0x1f, 0xd1, 0x1f, 0xf3, 0x9f,
+	0x40, 0x5b, 0x8a, 0x80, 0x47, 0xd2, 0x0d, 0xa6, 0x29, 0x60, 0xbd, 0x10, 0x28, 0xb4, 0xc6, 0xae,
+	0x08, 0x22, 0x6a, 0x40, 0xce, 0x69, 0xeb, 0xe4, 0x8c, 0x5c, 0x1c, 0xd5, 0x80, 0xa9, 0x90, 0x5e,
+	0xcf, 0x69, 0x77, 0x72, 0x4e, 0x9b, 0x9e, 0xc3, 0x41, 0x14, 0x7b, 0x8f, 0xbc, 0x2f, 0xf5, 0xdc,
+	0x3f, 0x20, 0x3b, 0xa2, 0xb7, 0x50, 0xe4, 0x4f, 0x53, 0x11, 0xf2, 0xa8, 0x21, 0xf5, 0x3d, 0xc5,
+	0x19, 0x2c, 0x6d, 0xc0, 0xb2, 0x06, 0xac, 0x9b, 0x35, 0x68, 0xee, 0xcf, 0x3f, 0x2a, 0xa4, 0xf3,
+	0x6b, 0xb1, 0x4e, 0x21, 0xdf, 0xc5, 0x11, 0x9f, 0xd0, 0x32, 0xe4, 0xe5, 0x6a, 0x50, 0x6d, 0x8a,
+	0x9d, 0xf4, 0xa7, 0xf6, 0x4a, 0xa0, 0xd4, 0x58, 0x5f, 0x1c, 0xbd, 0x82, 0x42, 0x2b, 0xe4, 0xae,
+	0xe4, 0xf4, 0x84, 0x6d, 0xee, 0x39, 0x7d, 0x93, 0x51, 0xfe, 0x23, 0xab, 0x78, 0x4b, 0xa3, 0xd7,
+	0x70, 0x78, 0xef, 0x8e, 0xc5, 0x60, 0x65, 0xdd, 0xca, 0x18, 0xdb, 0x03, 0x2d, 0x8d, 0xde, 0x00,
+	0x38, 0x93, 0x59, 0x66, 0xde, 0x71, 0xef, 0x2e, 0x77, 0x53, 0x7f, 0x5b, 0x9a, 0x64, 0xb1, 0x34,
+	0xc9, 0xe7, 0xd2, 0x24, 0xf3, 0xc4, 0xd4, 0x16, 0x89, 0xa9, 0xbd, 0x27, 0xa6, 0xe6, 0x15, 0xd4,
+	0x86, 0xea, 0x5f, 0x01, 0x00, 0x00, 0xff, 0xff, 0x05, 0x7d, 0x75, 0x12, 0x31, 0x02, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -205,9 +177,9 @@ const _ = grpc.SupportPackageIsVersion4
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
 type AuthorizationClient interface {
-	Create(ctx context.Context, in *TokenRequest, opts ...grpc.CallOption) (*Token, error)
-	Validate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*TokenResponse, error)
-	Invalidate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*TokenResponse, error)
+	Create(ctx context.Context, in *Claims, opts ...grpc.CallOption) (*Token, error)
+	Validate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*Claims, error)
+	Invalidate(ctx context.Context, in *Claims, opts ...grpc.CallOption) (*Claims, error)
 }
 
 type authorizationClient struct {
@@ -218,7 +190,7 @@ func NewAuthorizationClient(cc *grpc.ClientConn) AuthorizationClient {
 	return &authorizationClient{cc}
 }
 
-func (c *authorizationClient) Create(ctx context.Context, in *TokenRequest, opts ...grpc.CallOption) (*Token, error) {
+func (c *authorizationClient) Create(ctx context.Context, in *Claims, opts ...grpc.CallOption) (*Token, error) {
 	out := new(Token)
 	err := c.cc.Invoke(ctx, "/authorization.Authorization/Create", in, out, opts...)
 	if err != nil {
@@ -227,8 +199,8 @@ func (c *authorizationClient) Create(ctx context.Context, in *TokenRequest, opts
 	return out, nil
 }
 
-func (c *authorizationClient) Validate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*TokenResponse, error) {
-	out := new(TokenResponse)
+func (c *authorizationClient) Validate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*Claims, error) {
+	out := new(Claims)
 	err := c.cc.Invoke(ctx, "/authorization.Authorization/Validate", in, out, opts...)
 	if err != nil {
 		return nil, err
@@ -236,8 +208,8 @@ func (c *authorizationClient) Validate(ctx context.Context, in *Token, opts ...g
 	return out, nil
 }
 
-func (c *authorizationClient) Invalidate(ctx context.Context, in *Token, opts ...grpc.CallOption) (*TokenResponse, error) {
-	out := new(TokenResponse)
+func (c *authorizationClient) Invalidate(ctx context.Context, in *Claims, opts ...grpc.CallOption) (*Claims, error) {
+	out := new(Claims)
 	err := c.cc.Invoke(ctx, "/authorization.Authorization/Invalidate", in, out, opts...)
 	if err != nil {
 		return nil, err
@@ -247,9 +219,9 @@ func (c *authorizationClient) Invalidate(ctx context.Context, in *Token, opts ..
 
 // AuthorizationServer is the server API for Authorization service.
 type AuthorizationServer interface {
-	Create(context.Context, *TokenRequest) (*Token, error)
-	Validate(context.Context, *Token) (*TokenResponse, error)
-	Invalidate(context.Context, *Token) (*TokenResponse, error)
+	Create(context.Context, *Claims) (*Token, error)
+	Validate(context.Context, *Token) (*Claims, error)
+	Invalidate(context.Context, *Claims) (*Claims, error)
 }
 
 func RegisterAuthorizationServer(s *grpc.Server, srv AuthorizationServer) {
@@ -257,7 +229,7 @@ func RegisterAuthorizationServer(s *grpc.Server, srv AuthorizationServer) {
 }
 
 func _Authorization_Create_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TokenRequest)
+	in := new(Claims)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -269,7 +241,7 @@ func _Authorization_Create_Handler(srv interface{}, ctx context.Context, dec fun
 		FullMethod: "/authorization.Authorization/Create",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(AuthorizationServer).Create(ctx, req.(*TokenRequest))
+		return srv.(AuthorizationServer).Create(ctx, req.(*Claims))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -293,7 +265,7 @@ func _Authorization_Validate_Handler(srv interface{}, ctx context.Context, dec f
 }
 
 func _Authorization_Invalidate_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(Token)
+	in := new(Claims)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -305,7 +277,7 @@ func _Authorization_Invalidate_Handler(srv interface{}, ctx context.Context, dec
 		FullMethod: "/authorization.Authorization/Invalidate",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(AuthorizationServer).Invalidate(ctx, req.(*Token))
+		return srv.(AuthorizationServer).Invalidate(ctx, req.(*Claims))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -331,7 +303,7 @@ var _Authorization_serviceDesc = grpc.ServiceDesc{
 	Metadata: "authorization.proto",
 }
 
-func (m *TokenRequest) Marshal() (dAtA []byte, err error) {
+func (m *Claims) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
 	n, err := m.MarshalTo(dAtA)
@@ -341,20 +313,40 @@ func (m *TokenRequest) Marshal() (dAtA []byte, err error) {
 	return dAtA[:n], nil
 }
 
-func (m *TokenRequest) MarshalTo(dAtA []byte) (int, error) {
+func (m *Claims) MarshalTo(dAtA []byte) (int, error) {
 	var i int
 	_ = i
 	var l int
 	_ = l
-	if m.UserID != nil {
+	if m.ID != nil {
 		dAtA[i] = 0xa
 		i++
-		i = encodeVarintAuthorization(dAtA, i, uint64(m.UserID.Size()))
-		n1, err := m.UserID.MarshalTo(dAtA[i:])
+		i = encodeVarintAuthorization(dAtA, i, uint64(m.ID.Size()))
+		n1, err := m.ID.MarshalTo(dAtA[i:])
 		if err != nil {
 			return 0, err
 		}
 		i += n1
+	}
+	if m.Subject != nil {
+		dAtA[i] = 0x12
+		i++
+		i = encodeVarintAuthorization(dAtA, i, uint64(m.Subject.Size()))
+		n2, err := m.Subject.MarshalTo(dAtA[i:])
+		if err != nil {
+			return 0, err
+		}
+		i += n2
+	}
+	if m.ExpiresAt != nil {
+		dAtA[i] = 0x1a
+		i++
+		i = encodeVarintAuthorization(dAtA, i, uint64(github_com_gogo_protobuf_types.SizeOfStdTime(*m.ExpiresAt)))
+		n3, err := github_com_gogo_protobuf_types.StdTimeMarshalTo(*m.ExpiresAt, dAtA[i:])
+		if err != nil {
+			return 0, err
+		}
+		i += n3
 	}
 	return i, nil
 }
@@ -383,44 +375,6 @@ func (m *Token) MarshalTo(dAtA []byte) (int, error) {
 	return i, nil
 }
 
-func (m *TokenResponse) Marshal() (dAtA []byte, err error) {
-	size := m.Size()
-	dAtA = make([]byte, size)
-	n, err := m.MarshalTo(dAtA)
-	if err != nil {
-		return nil, err
-	}
-	return dAtA[:n], nil
-}
-
-func (m *TokenResponse) MarshalTo(dAtA []byte) (int, error) {
-	var i int
-	_ = i
-	var l int
-	_ = l
-	if m.Success {
-		dAtA[i] = 0x8
-		i++
-		if m.Success {
-			dAtA[i] = 1
-		} else {
-			dAtA[i] = 0
-		}
-		i++
-	}
-	if m.UserID != nil {
-		dAtA[i] = 0x12
-		i++
-		i = encodeVarintAuthorization(dAtA, i, uint64(m.UserID.Size()))
-		n2, err := m.UserID.MarshalTo(dAtA[i:])
-		if err != nil {
-			return 0, err
-		}
-		i += n2
-	}
-	return i, nil
-}
-
 func encodeVarintAuthorization(dAtA []byte, offset int, v uint64) int {
 	for v >= 1<<7 {
 		dAtA[offset] = uint8(v&0x7f | 0x80)
@@ -430,14 +384,22 @@ func encodeVarintAuthorization(dAtA []byte, offset int, v uint64) int {
 	dAtA[offset] = uint8(v)
 	return offset + 1
 }
-func (m *TokenRequest) Size() (n int) {
+func (m *Claims) Size() (n int) {
 	if m == nil {
 		return 0
 	}
 	var l int
 	_ = l
-	if m.UserID != nil {
-		l = m.UserID.Size()
+	if m.ID != nil {
+		l = m.ID.Size()
+		n += 1 + l + sovAuthorization(uint64(l))
+	}
+	if m.Subject != nil {
+		l = m.Subject.Size()
+		n += 1 + l + sovAuthorization(uint64(l))
+	}
+	if m.ExpiresAt != nil {
+		l = github_com_gogo_protobuf_types.SizeOfStdTime(*m.ExpiresAt)
 		n += 1 + l + sovAuthorization(uint64(l))
 	}
 	return n
@@ -456,22 +418,6 @@ func (m *Token) Size() (n int) {
 	return n
 }
 
-func (m *TokenResponse) Size() (n int) {
-	if m == nil {
-		return 0
-	}
-	var l int
-	_ = l
-	if m.Success {
-		n += 2
-	}
-	if m.UserID != nil {
-		l = m.UserID.Size()
-		n += 1 + l + sovAuthorization(uint64(l))
-	}
-	return n
-}
-
 func sovAuthorization(x uint64) (n int) {
 	for {
 		n++
@@ -485,7 +431,7 @@ func sovAuthorization(x uint64) (n int) {
 func sozAuthorization(x uint64) (n int) {
 	return sovAuthorization(uint64((x << 1) ^ uint64((int64(x) >> 63))))
 }
-func (m *TokenRequest) Unmarshal(dAtA []byte) error {
+func (m *Claims) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
 	iNdEx := 0
 	for iNdEx < l {
@@ -508,15 +454,15 @@ func (m *TokenRequest) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: TokenRequest: wiretype end group for non-group")
+			return fmt.Errorf("proto: Claims: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: TokenRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: Claims: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		case 1:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field UserID", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field ID", wireType)
 			}
 			var msglen int
 			for shift := uint(0); ; shift += 7 {
@@ -540,10 +486,76 @@ func (m *TokenRequest) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			if m.UserID == nil {
-				m.UserID = &uuid.UUID{}
+			if m.ID == nil {
+				m.ID = &uuid.UUID{}
 			}
-			if err := m.UserID.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+			if err := m.ID.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Subject", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuthorization
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= (int(b) & 0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthAuthorization
+			}
+			postIndex := iNdEx + msglen
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Subject == nil {
+				m.Subject = &uuid.UUID{}
+			}
+			if err := m.Subject.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ExpiresAt", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowAuthorization
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= (int(b) & 0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthAuthorization
+			}
+			postIndex := iNdEx + msglen
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.ExpiresAt == nil {
+				m.ExpiresAt = new(time.Time)
+			}
+			if err := github_com_gogo_protobuf_types.StdTimeUnmarshal(m.ExpiresAt, dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
@@ -625,109 +637,6 @@ func (m *Token) Unmarshal(dAtA []byte) error {
 				return io.ErrUnexpectedEOF
 			}
 			m.Token = string(dAtA[iNdEx:postIndex])
-			iNdEx = postIndex
-		default:
-			iNdEx = preIndex
-			skippy, err := skipAuthorization(dAtA[iNdEx:])
-			if err != nil {
-				return err
-			}
-			if skippy < 0 {
-				return ErrInvalidLengthAuthorization
-			}
-			if (iNdEx + skippy) > l {
-				return io.ErrUnexpectedEOF
-			}
-			iNdEx += skippy
-		}
-	}
-
-	if iNdEx > l {
-		return io.ErrUnexpectedEOF
-	}
-	return nil
-}
-func (m *TokenResponse) Unmarshal(dAtA []byte) error {
-	l := len(dAtA)
-	iNdEx := 0
-	for iNdEx < l {
-		preIndex := iNdEx
-		var wire uint64
-		for shift := uint(0); ; shift += 7 {
-			if shift >= 64 {
-				return ErrIntOverflowAuthorization
-			}
-			if iNdEx >= l {
-				return io.ErrUnexpectedEOF
-			}
-			b := dAtA[iNdEx]
-			iNdEx++
-			wire |= (uint64(b) & 0x7F) << shift
-			if b < 0x80 {
-				break
-			}
-		}
-		fieldNum := int32(wire >> 3)
-		wireType := int(wire & 0x7)
-		if wireType == 4 {
-			return fmt.Errorf("proto: TokenResponse: wiretype end group for non-group")
-		}
-		if fieldNum <= 0 {
-			return fmt.Errorf("proto: TokenResponse: illegal tag %d (wire type %d)", fieldNum, wire)
-		}
-		switch fieldNum {
-		case 1:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Success", wireType)
-			}
-			var v int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowAuthorization
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				v |= (int(b) & 0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			m.Success = bool(v != 0)
-		case 2:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field UserID", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowAuthorization
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= (int(b) & 0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthAuthorization
-			}
-			postIndex := iNdEx + msglen
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			if m.UserID == nil {
-				m.UserID = &uuid.UUID{}
-			}
-			if err := m.UserID.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex