+++ api/authorization.pb.micro.go
@@ -5,7 +5,9 @@ package authorization
 
 import (
 	fmt "fmt"
+	_ "github.com/gogo/protobuf/gogoproto"
 	proto "github.com/golang/protobuf/proto"
+	_ "github.com/golang/protobuf/ptypes/timestamp"
 	_ "github.com/koverto/uuid"
 	math "math"
 )
@@ -35,9 +37,9 @@ var _ server.Option
 // Client API for Authorization service
 
 type AuthorizationService interface {
-	Create(ctx context.Context, in *TokenRequest, opts ...client.CallOption) (*Token, error)
-	Validate(ctx context.Context, in *Token, opts ...client.CallOption) (*TokenResponse, error)
-	Invalidate(ctx context.Context, in *Token, opts ...client.CallOption) (*TokenResponse, error)
+	Create(ctx context.Context, in *Claims, opts ...client.CallOption) (*Token, error)
+	Validate(ctx context.Context, in *Token, opts ...client.CallOption) (*Claims, error)
+	Invalidate(ctx context.Context, in *Claims, opts ...client.CallOption) (*Claims, error)
 }
 
 type authorizationService struct {
@@ -52,7 +54,7 @@ func NewAuthorizationService(name string, c client.Client) AuthorizationService 
 	}
 }
 
-func (c *authorizationService) Create(ctx context.Context, in *TokenRequest, opts ...client.CallOption) (*Token, error) {
+func (c *authorizationService) Create(ctx context.Context, in *Claims, opts ...client.CallOption) (*Token, error) {
 	req := c.c.NewRequest(c.name, "Authorization.Create", in)
 	out := new(Token)
 	err := c.c.Call(ctx, req, out, opts...)
@@ -62,9 +64,9 @@ func (c *authorizationService) Create(ctx context.Context, in *TokenRequest, opt
 	return out, nil
 }
 
-func (c *authorizationService) Validate(ctx context.Context, in *Token, opts ...client.CallOption) (*TokenResponse, error) {
+func (c *authorizationService) Validate(ctx context.Context, in *Token, opts ...client.CallOption) (*Claims, error) {
 	req := c.c.NewRequest(c.name, "Authorization.Validate", in)
-	out := new(TokenResponse)
+	out := new(Claims)
 	err := c.c.Call(ctx, req, out, opts...)
 	if err != nil {
 		return nil, err
@@ -72,9 +74,9 @@ func (c *authorizationService) Validate(ctx context.Context, in *Token, opts ...
 	return out, nil
 }
 
-func (c *authorizationService) Invalidate(ctx context.Context, in *Token, opts ...client.CallOption) (*TokenResponse, error) {
+func (c *authorizationService) Invalidate(ctx context.Context, in *Claims, opts ...client.CallOption) (*Claims, error) {
 	req := c.c.NewRequest(c.name, "Authorization.Invalidate", in)
-	out := new(TokenResponse)
+	out := new(Claims)
 	err := c.c.Call(ctx, req, out, opts...)
 	if err != nil {
 		return nil, err
@@ -85,16 +87,16 @@ func (c *authorizationService) Invalidate(ctx context.Context, in *Token, opts .
 // Server API for Authorization service
 
 type AuthorizationHandler interface {
-	Create(context.Context, *TokenRequest, *Token) error
-	Validate(context.Context, *Token, *TokenResponse) error
-	Invalidate(context.Context, *Token, *TokenResponse) error
+	Create(context.Context, *Claims, *Token) error
+	Validate(context.Context, *Token, *Claims) error
+	Invalidate(context.Context, *Claims, *Claims) error
 }
 
 func RegisterAuthorizationHandler(s server.Server, hdlr AuthorizationHandler, opts ...server.HandlerOption) error {
 	type authorization interface {
-		Create(ctx context.Context, in *TokenRequest, out *Token) error
-		Validate(ctx context.Context, in *Token, out *TokenResponse) error
-		Invalidate(ctx context.Context, in *Token, out *TokenResponse) error
+		Create(ctx context.Context, in *Claims, out *Token) error
+		Validate(ctx context.Context, in *Token, out *Claims) error
+		Invalidate(ctx context.Context, in *Claims, out *Claims) error
 	}
 	type Authorization struct {
 		authorization
@@ -107,14 +109,14 @@ type authorizationHandler struct {
 	AuthorizationHandler
 }
 
-func (h *authorizationHandler) Create(ctx context.Context, in *TokenRequest, out *Token) error {
+func (h *authorizationHandler) Create(ctx context.Context, in *Claims, out *Token) error {
 	return h.AuthorizationHandler.Create(ctx, in, out)
 }
 
-func (h *authorizationHandler) Validate(ctx context.Context, in *Token, out *TokenResponse) error {
+func (h *authorizationHandler) Validate(ctx context.Context, in *Token, out *Claims) error {
 	return h.AuthorizationHandler.Validate(ctx, in, out)
 }
 
-func (h *authorizationHandler) Invalidate(ctx context.Context, in *Token, out *TokenResponse) error {
+func (h *authorizationHandler) Invalidate(ctx context.Context, in *Claims, out *Claims) error {
 	return h.AuthorizationHandler.Invalidate(ctx, in, out)
 }